# server/swagger.yaml
openapi: 3.0.0
info:
  title: Shopping Online API
  description: API documentation for the Shopping Online project.
  version: 1.0.0
  contact:
    name: Admin Team

servers:
  - url: http://127.0.0.1:3013
    description: Development Server

paths:
  /api/products:
    get:
      summary: Get all products
      description: Returns all products from the public product API.
      tags:
        - Public Products
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          description: Server error
    post:
      summary: Create a product
      description: Creates a new product in the public product API.
      tags:
        - Public Products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid request body
        '500':
          description: Server error

  /api/products/{id}:
    put:
      summary: Update a product
      description: Updates a product in the public product API.
      tags:
        - Public Products
      parameters:
        - name: id
          in: path
          required: true
          description: Product id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
        '500':
          description: Server error
    delete:
      summary: Delete a product
      description: Deletes a product in the public product API.
      tags:
        - Public Products
      parameters:
        - name: id
          in: path
          required: true
          description: Product id
          schema:
            type: string
      responses:
        '200':
          description: Product deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  product:
                    $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
        '500':
          description: Server error

  /api/admin/login:
    post:
      summary: Admin login
      description: Authenticates an admin account and returns a JWT token.
      tags:
        - Admin Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginInput'
      responses:
        '200':
          description: Login result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminLoginResponse'
        '500':
          description: Server error

  /api/admin/orders:
    get:
      summary: Get all orders
      description: Returns all orders sorted by creation date in descending order.
      tags:
        - Admin Orders
      parameters:
        - $ref: '#/components/parameters/XAccessToken'
      security:
        - AdminTokenAuth: []
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          description: Missing or invalid admin token
        '500':
          description: Server error

  /api/admin/orders/customer/{cid}:
    get:
      summary: Get orders by customer id
      description: Returns all orders for a specific customer.
      tags:
        - Admin Customers
      parameters:
        - name: cid
          in: path
          required: true
          description: Customer id
          schema:
            type: string
        - $ref: '#/components/parameters/XAccessToken'
      security:
        - AdminTokenAuth: []
      responses:
        '200':
          description: List of orders for the customer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          description: Missing or invalid admin token
        '500':
          description: Server error

  /api/admin/orders/status/{id}:
    put:
      summary: Update order status
      description: Updates the status of an order by id.
      tags:
        - Admin Orders
      parameters:
        - name: id
          in: path
          required: true
          description: Order id
          schema:
            type: string
        - $ref: '#/components/parameters/XAccessToken'
      security:
        - AdminTokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderStatusInput'
      responses:
        '200':
          description: Updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          description: Missing or invalid admin token
        '500':
          description: Server error

  /api/admin/customers:
    get:
      summary: Get all customers
      description: Returns all customers for the admin customer screen.
      tags:
        - Admin Customers
      parameters:
        - $ref: '#/components/parameters/XAccessToken'
      security:
        - AdminTokenAuth: []
      responses:
        '200':
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerSummary'
        '401':
          description: Missing or invalid admin token
        '500':
          description: Server error

  /api/admin/customers/deactive/{id}:
    put:
      summary: Deactivate a customer
      description: Sets the specified customer to inactive using the customer's token.
      tags:
        - Admin Customers
      parameters:
        - name: id
          in: path
          required: true
          description: Customer id
          schema:
            type: string
        - $ref: '#/components/parameters/XAccessToken'
      security:
        - AdminTokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerDeactivateInput'
      responses:
        '200':
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerSummary'
        '401':
          description: Missing or invalid admin token
        '500':
          description: Server error

  /api/admin/customers/sendmail/{id}:
    get:
      summary: Send activation email to a customer
      description: Sends the signup verification email again to the specified customer.
      tags:
        - Admin Customers
      parameters:
        - name: id
          in: path
          required: true
          description: Customer id
          schema:
            type: string
        - $ref: '#/components/parameters/XAccessToken'
      security:
        - AdminTokenAuth: []
      responses:
        '200':
          description: Email sending result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendEmailResponse'
        '401':
          description: Missing or invalid admin token
        '500':
          description: Server error

  /api/customer/categories:
    get:
      summary: Get categories
      description: Returns all product categories.
      tags:
        - Customer Data
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategorySummary'
        '500':
          description: Server error

  /api/customer/products/new:
    get:
      summary: Get newest products
      description: Returns the 3 newest products.
      tags:
        - Customer Data
      responses:
        '200':
          description: List of newest products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          description: Server error

  /api/customer/products/hot:
    get:
      summary: Get hottest products
      description: Returns the 3 hottest approved products.
      tags:
        - Customer Data
      responses:
        '200':
          description: List of hottest products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          description: Server error

  /api/customer/products/category/{cid}:
    get:
      summary: Get products by category
      description: Returns all products in the specified category.
      tags:
        - Customer Data
      parameters:
        - name: cid
          in: path
          required: true
          description: Category id
          schema:
            type: string
      responses:
        '200':
          description: List of products in the category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          description: Server error

  /api/customer/products/search/{keyword}:
    get:
      summary: Search products
      description: Searches products by keyword.
      tags:
        - Customer Data
      parameters:
        - name: keyword
          in: path
          required: true
          description: Search keyword
          schema:
            type: string
      responses:
        '200':
          description: Matching products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          description: Server error

  /api/customer/products/{id}:
    get:
      summary: Get product by id
      description: Returns product details by id.
      tags:
        - Customer Data
      parameters:
        - name: id
          in: path
          required: true
          description: Product id
          schema:
            type: string
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
        '500':
          description: Server error

components:
  securitySchemes:
    AdminTokenAuth:
      type: apiKey
      in: header
      name: x-access-token
  parameters:
    XAccessToken:
      name: x-access-token
      in: header
      required: true
      description: Admin JWT token returned by POST /api/admin/login
      schema:
        type: string
  schemas:
    AdminLoginInput:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: thanh
        password:
          type: string
          example: '123'

    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        token:
          type: string

    CategorySummary:
      type: object
      properties:
        _id:
          type: string
          description: Category id
        name:
          type: string
          description: Category name

    CustomerSummary:
      type: object
      properties:
        _id:
          type: string
          description: Customer id
        username:
          type: string
        password:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        active:
          type: number
        token:
          type: string

    CustomerDeactivateInput:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: Customer token used to deactivate the account

    SendEmailResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        preview:
          type: object
          properties:
            email:
              type: string
            id:
              type: string
            token:
              type: string
            note:
              type: string

    Product:
      type: object
      properties:
        _id:
          type: string
          description: Product id
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        cdate:
          type: number
          description: Product created time in milliseconds
        category:
          $ref: '#/components/schemas/CategorySummary'

    ProductInput:
      type: object
      required:
        - name
        - description
        - price
        - imageUrl
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
          format: uri

    OrderItem:
      type: object
      properties:
        product:
          $ref: '#/components/schemas/Product'
        quantity:
          type: number

    Order:
      type: object
      properties:
        _id:
          type: string
          description: Order id
        cdate:
          type: number
          description: Order created time in milliseconds
        total:
          type: number
        status:
          type: string
        customer:
          $ref: '#/components/schemas/CustomerSummary'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'

    OrderStatusInput:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - PENDING
            - APPROVED
            - CANCELED

tags:
  - name: Public Products
    description: Public CRUD endpoints defined in product routes.
  - name: Customer Data
    description: Customer-facing endpoints for categories and products.
  - name: Admin Auth
    description: Admin authentication endpoints.
  - name: Admin Orders
    description: Admin order management endpoints.
  - name: Admin Customers
    description: Admin customer list and customer order lookup endpoints.